+++ src/Common/Tests/nanosleepTest.c
@@ -46,7 +46,8 @@ int main()
    vdscEndTimer( &timer );
    vdscCalculateTimer( &timer, &sec, &nanoSec );
 
-   fprintf( stderr, "Sec = %u, uSec = %u\n", sec, nanoSec/1000 );
+   fprintf( stderr, "Sec = %u, uSec = %u\n", (unsigned int)sec, 
+                                             (unsigned int)(nanoSec/1000) );
    if ( sec == 0 && nanoSec < 10000 )
    {
       fprintf( stderr, "Timer returns invalid time!\n" );
